+++ bricks/flutter_starter_brick/__brick__/lib/dependency_injection/src/register_instances.dart
@@ -33,7 +33,6 @@ void registerInstances(GetIt getIt, {required Environment environment}) {
     ..registerLazySingleton<EnvironmentVariables>(
       () => switch (getIt.get<Environment>()) {
         Environment.development => const DevelopmentEnvironmentVariables(),
-        Environment.staging => const StagingEnvironmentVariables(),
         Environment.production => const ProductionEnvironmentVariables(),
       },
     )