+++ bricks/flutter_starter_brick/__brick__/lib/foundation/environment_variables/src/environment_variables.dart
@@ -20,17 +20,6 @@ final class DevelopmentEnvironmentVariables extends EnvironmentVariables {
   }
 }
 
-/// Staging environment variables.
-final class StagingEnvironmentVariables extends EnvironmentVariables {
-  const StagingEnvironmentVariables() : super._();
-
-  @override
-  String get backendBaseUrl {
-    // TODO({{dev_name}}): Update with your staging API URL
-    return 'https://api.chucknorris.io';
-  }
-}
-
 /// Production environment variables.
 final class ProductionEnvironmentVariables extends EnvironmentVariables {
   const ProductionEnvironmentVariables() : super._();