+++ bricks/flutter_starter_brick/__brick__/lib/main_staging.dart
@@ -1,7 +1,0 @@
-import 'package:{{proj_name}}/foundation/environments/environments.dart';
-import 'package:{{proj_name}}/main_common.dart';
-
-Future<void> main() async {
-  // run the app with the staging environment
-  await mainCommon(Environment.staging);
-}